Conflicts:
	src/com/luxoft/bankapp/application/BankApplication.java
	src/com/luxoft/bankapp/application/BankClientRemoteOffice.java
	src/com/luxoft/bankapp/application/BankServer.java
	src/com/luxoft/bankapp/application/BankUI.java
	src/com/luxoft/bankapp/commander/Commander.java
	src/com/luxoft/bankapp/commander/command/BankInfoCommand.java
	src/com/luxoft/bankapp/commander/command/FindClientCommand.java
	src/com/luxoft/bankapp/commander/command/GetActiveAccountCommand.java
	src/com/luxoft/bankapp/commander/command/LoadCommand.java
	src/com/luxoft/bankapp/commander/command/SaveCommand.java
	src/com/luxoft/bankapp/commander/command/WithdrawCommand.java
	src/com/luxoft/bankapp/commander/command/servicecommands/AddAccountCommand.java
	src/com/luxoft/bankapp/commander/command/servicecommands/AddClientCommand.java
	src/com/luxoft/bankapp/commander/command/servicecommands/BankFeedCommand.java
	src/com/luxoft/bankapp/commander/command/servicecommands/DepositCommand.java
	src/com/luxoft/bankapp/commander/command/servicecommands/SetActiveAccountCommand.java
	src/com/luxoft/bankapp/commander/command/servicecommands/TransferCommand.java
	src/com/luxoft/bankapp/model/impl/Bank.java
	src/com/luxoft/bankapp/model/impl/Client.java

+++ src/com/luxoft/bankapp/dao/ClientDao.java
@@ -16,7 +16,7 @@ import java.util.List;
 public interface ClientDao {
     Client findClientByName(Bank bank, String name) throws ClientNotExistsException, DaoException;
 
-    Client findClientById(int id) throws ClientNotExistsException, DaoException;
+    Client findClientById(long id) throws ClientNotExistsException, DaoException;
 
     List<Client> getAllClients(Bank bank) throws DaoException;
